requirements.txt
flask
openai
python-dotenv
pymongo
bcrypt
google-genai
pillow
pytest>=7.0.0
